fix: tint hair and skin
